# Add compact entry points for ARM