This pull request adds compact entry points for ARM. 

Related issue: #11076 

For MethodDescChunk with 5 MethodDescs compact entry points will look like this:

```
entry1:
  mov r12, pc
  b CentralJump
entry2:
  mov r12, pc
  b CentralJump
entry3:
  mov r12, pc
  b CentralJump
entry4:
  mov r12, pc
  b CentralJump
entry5:
  mov r12, pc
  b CentralJump
CentraJump:
  ldr pc, pThePreStubCompactARM
  nop
  dw pChunk
  dw pThePreStubCompactARM
```

Compact entry points are encoded in Thumb, each compact entry has 4 byte size, CentralJump is 14 bytes.

ThePreStubCompactARM finds CentralJump by looking into branch offset of compact entry. Then it calculates offset of the compact entry from CentralJump by program counter stored in r12 and, thus, calculates index of MethodDesc in MethodDescChunk. Then,ThePreStubCompactARM loads pointer to MethodDescChunk from CentralJump and obtains MethodDesc, corresponding to the calculated index (this is done by traversing all MethodDescs in MethodDescChunk starting from the first until the MethodDesc with required index is met). After that, ThePreStubCompactARM calls PreStubWorker with obtained MethodDesc, like original ThePreStub.

Estimation on Tizen GUI Xamarin application (https://developer.tizen.org/sites/default/files/documentation/puzzle2.zip) shows, for ReadyToRun mode, that reduction of code heap size is 76 KB, which is 20% of code heap and, approximately, 2.2% of overall per-application CoreCLR memory consumption. For No-Precompilation mode, reduction of code heap size is 216 KB, which is 28% of code heap and, approximately, 2.3% of overall per-application CoreCLR memory consumption.

ReadyToRun mode means the Tizen-default set of precompiled assemblies is in ReadyToRun format, No-Precompilation mode means that there are no precompiled assemblies.

Testing:

- CI: default
- CoreCLR test suite on ARM on Tizen: no regression happened 

Micro-benchmark, provided by @jkotas in https://github.com/dotnet/coreclr/issues/11076#issuecomment-296748946, showed next results:
- Cortex-A7 softfp:
      before: 12.81s (user time)
      after: 10.488s (user time)
- Cortex-A15 softfp:
      before: 5.3s (user time)
      after: 4.68s (user time)

@Dmitri-Botcharnikov @ruben-ayrapetyan