@gbalykov 2017-04-20
For ReadyToRun mode memory reduction is lower than for No-Precompilation mode. Could you, please, suggest, why this is happening?