@gbalykov 2017-04-27
@dotnet-bot test Tizen armel Cross Debug Build