@gbalykov 2017-04-27
Test JIT/Regression/CLR-x86-JIT/V1-M09.5-PDC/b26863/b26863/b26863.sh in CI for Tizen failed and succeeded in two consecutive launches without any additional fixes

fail -> https://ci.dot.net/job/dotnet_coreclr/job/master/job/armel_cross_debug_tizen_prtest/1757/console
success -> https://ci.dot.net/job/dotnet_coreclr/job/master/job/armel_cross_debug_tizen_prtest/1758/console

What do you think about this? Could this be related to this pull request or this is a known issue in CI?