@danmosemsft 2017-05-05
@jkotas for question above? although CI is green, should this merge?